Update documentation and improve memory management in physics engine
- Added CDN usage instructions to README for easier integration.
- Updated default max neighbors for spatial configuration to 128, with guidance for dense flocks.
- Enhanced memory structure documentation to reflect changes in entity storage and buffer sizes.
- Introduced new metrics for heap usage in Box2D, including high-water marks and memory insights.
- Adjusted various components to utilize Uint16 for entity indices, optimizing memory usage.
- Improved the MemoryPanel to display detailed capacity insights and heap statistics.
- Refined query system to dynamically size query results based on entity capacity, enhancing performance.

+++ src/components/RigidBody.js
@@ -47,7 +47,7 @@ export class RigidBody extends Component {
 
     // Sleeping optimization (gated by physics.sleeping config)
     sleeping: Uint8Array, // 0 = awake, 1 = sleeping
-    stillnessTime: Float32Array, // Consecutive still frames (particle ticks), not seconds
+    stillnessTime: Uint16Array, // Consecutive still frames (particle ticks), not seconds
   };
 
   // ═══════════════════════════════════════════════════════════════════════════
